requirements.txt
openai
groq
requests
python-dotenv
